fixed content edifor for contetn blocks in pages
