+++ app/components/ui/form.input.urls.tsx
@@ -43,8 +43,6 @@ export default function URLsInput({
   const [urls, setUrls] = useState<URL[]>(value || []);
 
   useEffect(() => {
-    console.log(value);
-
     if (!!value.length) {
       setUrls(value || []);
     }
